QA-15380: Issue with customized preview and aliased user (#18)
Fix an issue with customized preview for Javascript modules for content with virtual nodes.

The virtual nodes are correctly created (temporarily) in the session of the aliased user, however the rendering was performed with the current user, causing a ItemNotFoundException. The rendering now uses the aliased user if any, and fallback to the current user otherwise.