feat: Add priority to jahiaComponent (#213)
Add a new priority parameter when defining components (both views and templates) to decide which component should be used when rendering a given view/template, in case several components match a given resource.
The parameter is optional and its default value is 0. It can be set to a negative or a positive number.
The component with the highest priority has precedence over the other components.

NB: the identifiers of the registered components now include the priority, only if the priority is different from the default value (0), to remain backward compatible with JS modules that might have been deployed with a prior version of Javascript Modules Engine.