feat: Make Hydrogen a sample test module (#217)
Add a test module by following the Getting Started tutorial.
The related Cypress tests are broken down into steps, to follow the tutorial's approach.

Before starting the tests, the sample tutorial site is created, as the result of the Getting Started tutorial.
The creation is done by importing the site from the hydrogen-prepackaged.
For "regular" Github workflows (On Code Change and On merge to main), the file artifacts/javascript-modules-samples-hydrogen-prepackaged-*.jar is used, generated during the build and moved to the ./artifacts/ folder.
For nightly/manual tests, the prepackaged site is fetched from a Maven repository.
To support both use cases, an environment variable CYPRESS_PREPACKAGED_SITE_URL is set in the Github actions to differentiate those 2 scenarios.